dev-assistant-endpoint-server/requirements.txt
uvicorn~=0.22.0
fastapi~=0.95.1
transformers
accelerate>=0.20.3
six
marko
mysql-connector-python
gradio